pyproject.toml
[tool.poetry]
name = "xgeeks Backend"
version = "0.1.0"
description = ""
authors = ["Miguel Jesus <miguelangeloleal@hotmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
python-dotenv = "0.18.0"
psycopg2-binary = "2.9.1"
SQLAlchemy = "1.4.25"
uvicorn = "0.15"
pydantic = "1.8.2"
asyncpg = "0.24.0"
databases = "0.5.3"
fastapi = "0.70.0"
FastAPI-SQLAlchemy = "0.2.1"
PyJWT = "2.3.0"
rsa = "4.7.2"
fastapi_jwt_auth = { git = "https://github.com/wygame-io/fastapi-jwt-auth.git", tag = "v0.6.1" }
passlib = "^1.7.4"
envyaml = "*"
python-multipart = "^0.0.5"
python-dateutil = "^2.8.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
